refactor: 메뉴 패키지 구조 변경 menuGroup과 menu를 하나의 컨텍스트로 바라보았다
