+++ src/main/java/kitchenpos/tobe/menu/domain/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.tobe.menu.domain;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
