+++ src/main/java/kitchenpos/tobe/menu/domain/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.tobe.menu.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -10,6 +10,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
+import kitchenpos.tobe.product.domain.Product;
 
 import java.util.UUID;
