+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -1,12 +1,13 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroupRepository;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.MenuRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
-import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.tobe.menu.application.service.MenuService;
+import kitchenpos.tobe.menu.domain.Menu;
+import kitchenpos.tobe.menu.application.port.out.MenuGroupRepository;
+import kitchenpos.tobe.menu.domain.MenuProduct;
+import kitchenpos.tobe.menu.application.port.out.MenuRepository;
+import kitchenpos.tobe.product.domain.Product;
+import kitchenpos.tobe.product.application.port.out.ProductRepository;
+import kitchenpos.tobe.external.port.out.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
